+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ## Version X.Y.Z (YYYY-MM-DD)
 
 - Add PDF/A-3b compliant PDF file export
+- Add parallel job execution capability
+- Fix the `status` field of the `GET /status` request: The status will no longer be `IDLE` while processing exactly one job.
+- Changed response of `GET /status` route
+  - Adding field `jobs_processing`: Array containing the IDs of all the jobs currently being processed
+  - Adding field `jobs_queued`: Array containing the IDs of all the jobs currently in the queue (in order)
+  - Adding field `jobs_max`: Number of maximal jobs that can be run in parallel
+  - Adding field `queue_max`: Number of maximal queue size
 - Update Python dependencies
 - Automatically build `linux/amd64` and `linux/arm64` variants of the Docker image
 - Update GitHub actions workflow for Docker image build and release